GPUToolbox.jl

API

GPUToolbox.LazyInitialized — Type

LazyInitialized{T}()

A thread-safe, lazily-initialized wrapper for a value of type T. Initialize and fetch the value by calling get!. The constructor is ensured to only be called once.

This type is intended for lazy initialization of e.g. global structures, without using __init__. It is similar to protecting accesses using a lock, but is much cheaper.

GPUToolbox.Literal — Type

Literal{T}

Construct an object that can be used to convert literals to other types. One can use the object in suffix form 1*i8 or 1i8 to perform the conversion.

Exported constants

i8: Convert to Int8

i16: Convert to Int16

i32: Convert to Int32

u8: Convert to UInt8

u16: Convert to UInt16

u32: Convert to UInt32

GPUToolbox.SimpleVersion — Type

SimpleVersion(major, [minor])
SimpleVersion(v::AbstractString)

A GPU-compatible version number.

GPUToolbox.@checked — Macro

@checked function foo(...)
    rv = ...
    return rv
end

Macro for wrapping a function definition returning a status code. Two versions of the function will be generated: foo, with the function execution wrapped by an invocation of the check function (to be implemented by the caller of this macro), and unchecked_foo where no such invocation is present and the status code is returned to the caller.

GPUToolbox.@debug_ccall — Macro

@debug_ccall ...

A version of ccall` that prints the ccall, its arguments, and its return value.

GPUToolbox.@gcsafe_ccall — Macro

@gcsafe_ccall ...

Call a foreign function just like @ccall, but marking it safe for the GC to run. This is useful for functions that may block, so that the GC isn't blocked from running, but may also be required to prevent deadlocks (see JuliaGPU/CUDA.jl#2261).

Note that this is generally only safe with non-Julia C functions that do not call back into the Julia directly.

GPUToolbox.@memoize — Macro

@memoize [key::T] [maxlen=...] begin
    # expensive computation
end::T

Low-level, no-frills memoization macro that stores values in a thread-local, typed cache. The types of the caches are derived from the syntactical type assertions.

The cache consists of two levels, the outer one indexed with the thread index. If no key is specified, the second level of the cache is dropped.

If the the maxlen option is specified, the key is assumed to be an integer, and the secondary cache will be a vector with length maxlen. Otherwise, a dictionary is used.

GPUToolbox.@sv_str — Macro

sv"str"

Construct a SimpleVersion from str.